+++ frontend/src/hooks/graph/useGraphHover.ts
@@ -1,7 +1,7 @@
 import * as d3 from 'd3';
 import { useEffect } from 'react';
 
-export default function useGraphHover(nodeSelector: SVGGElement | null, nodes: any, hoveredNode: string) {
+export default function useGraphHover(nodeSelector: SVGGElement | null, nodes: any[], hoveredNode: string) {
   useEffect(() => {
     if (nodeSelector === null) return;
 
@@ -12,7 +12,7 @@ export default function useGraphHover(nodeSelector: SVGGElement | null, nodes: a
     d3.select(nodeSelector)
       .selectAll('text')
       .data(nodes)
-      .filter((d: any) => d.key === hoveredNode)
+      .filter((d) => d.key === hoveredNode)
       .style('fill-opacity', '1');
   }, [nodeSelector, hoveredNode, nodes]);
 }
